+++ src/components/Button.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
-export class Button extends React.Component {
-    public render(): JSX.Element { return <button></button>; }
+interface ButtonProps { kind: string }
+
+export class Button extends React.Component <ButtonProps, {}> {
+    public render(): JSX.Element { return <button>{this.props.kind}</button>; }
 }