+++ src/components/Timer.tsx
@@ -1,8 +1,16 @@
 import React from 'react';
 import { Button } from './Button';
 
-export const Timer = (): JSX.Element => 
-  <div>
-    Timer
-    <Button />
-  </div>;
+interface TimerState { time: string }
+
+export class Timer extends React.Component <{}, TimerState> {
+  public state: TimerState = { time: '00:00:00' };
+
+  public render(): JSX.Element { return (
+      <div>
+        {this.state.time}
+        <Button kind='test' />
+      </div>
+    );
+  }
+}